Add a varlink server and client for testing
Add a small prototype to test passing FDs via varlink and opening them
on the varlink server side.

The varlink server then uses libcryptsetup-rs to create the dm-verity
data for a given disk image that is passed in via varlink.

Currently, a patched libcryptsetup-rs (with
https://github.com/stratis-storage/libcryptsetup-rs/pull/461 applied)
is necessary for the dm-verity part to work.

This is also useful to test the varlink server in a locked-down
environment, as that's what we'll be working with as an actual sysupdate
pull backend later. For a systemd-run command to lock down the server,
see test_pull_server.rs.